backend/requirements.txt
# List your backend dependencies here
fastapi
uvicorn
python-multipart
google-generativeai
PyPDF2
jinja2
crawl4ai
pdf2image
pillow
